backend/requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
python-dotenv==1.0.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
pydantic==1.10.7
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
